fix: one wagmi provider
