+++ apps/web/src/pages/ido/[ido].tsx
@@ -14,4 +14,6 @@ CurrentIdoPage.chains = IDO_SUPPORT_CHAINS
 
 CurrentIdoPage.mp = true
 
+CurrentIdoPage.w3wWagmiConfig = true
+
 export default CurrentIdoPage
