+++ apps/web/src/utils/wagmi.ts
@@ -4,7 +4,7 @@ import { blocto } from '@pancakeswap/wagmi/connectors/blocto'
 import { CHAINS } from 'config/chains'
 import { PUBLIC_NODES } from 'config/nodes'
 import memoize from 'lodash/memoize'
-import { Transport } from 'viem'
+import { Transport, custom } from 'viem'
 import { createConfig, fallback, http } from 'wagmi'
 import { mainnet } from 'wagmi/chains'
 import { coinbaseWallet, injected, safe, walletConnect } from 'wagmi/connectors'
@@ -71,6 +71,27 @@ export const transports = chains.reduce((ts, chain) => {
   }
 }, {} as Record<number, Transport>)
 
+const injectedTransports = chains.reduce((ts, chain) => {
+  let httpStrings: string[] | readonly string[] = []
+
+  httpStrings = PUBLIC_NODES[chain.id] ? PUBLIC_NODES[chain.id] : []
+
+  const injectedTransport =
+    typeof window !== 'undefined' && window.ethereum ? custom(window.ethereum as any) : undefined
+
+  const allTransports = [injectedTransport, ...httpStrings.map((t: any) => http(t))].filter(Boolean) as Transport[]
+
+  if (ts) {
+    // eslint-disable-next-line no-param-reassign
+    ts[chain.id] = fallback(allTransports)
+    return ts
+  }
+
+  return {
+    [chain.id]: fallback(allTransports),
+  }
+}, {} as Record<number, Transport>)
+
 export const cyberWalletConnector = isCyberWallet()
   ? createCyberWalletConnector({
       name: 'PancakeSwap',
@@ -99,6 +120,18 @@ export function createWagmiConfig() {
 
       ...(cyberWalletConnector ? [cyberWalletConnector as any] : []),
     ],
+  })
+}
+
+export const createW3WWagmiConfig = () => {
+  return createConfig({
+    chains,
+    ssr: true,
+    syncConnectedChain: true,
+    transports: injectedTransports,
+    ...CLIENT_CONFIG,
+
+    connectors: [injectedConnector, walletConnectConnector, binanceWeb3WalletConnector()],
   })
 }
